+++ packages/plugins/presets/comfyui-media-provider/src/index.ts
@@ -1,0 +1,9 @@
+import { definePlugin } from "@vetta-org/plugin-sdk";
+import { createComfyUiProvider } from "./provider";
+
+export default definePlugin({
+	activate(ctx) {
+		ctx.media.registerProvider(createComfyUiProvider(ctx));
+	},
+});
+